// ui/src/lib/api/model/internalDeploymentsDeployment.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { InternalDeploymentsDeploymentRun } from './internalDeploymentsDeploymentRun';
import type { Runtime } from './runtime';
import type { DeploymentStatus } from './deploymentStatus';

/**
 * Deployment package with configuration and metadata
 */
export interface InternalDeploymentsDeployment {
  config?: number[];
  created_at?: string;
  description?: string;
  failed_runs?: number;
  id?: string;
  last_run_at?: string;
  name?: string;
  package_hash?: string;
  package_url?: string;
  /** Relationships */
  runs?: InternalDeploymentsDeploymentRun[];
  runtime?: Runtime;
  status?: DeploymentStatus;
  successful_runs?: number;
  /** Computed fields */
  total_runs?: number;
  updated_at?: string;
  version?: string;
}