// ui/src/lib/api/model/internalDeploymentsCreateDeploymentRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { InternalDeploymentsDeploymentConfig } from './internalDeploymentsDeploymentConfig';
import type { Runtime } from './runtime';

export interface InternalDeploymentsCreateDeploymentRequest {
  config?: InternalDeploymentsDeploymentConfig;
  description?: string;
  name: string;
  package_hash: string;
  package_url: string;
  runtime: Runtime;
  version: string;
}