// ui/src/lib/api/model/internalDeploymentsUpdateDeploymentRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { InternalDeploymentsDeploymentConfig } from './internalDeploymentsDeploymentConfig';
import type { DeploymentStatus } from './deploymentStatus';

export interface InternalDeploymentsUpdateDeploymentRequest {
  config?: InternalDeploymentsDeploymentConfig;
  description?: string;
  status?: DeploymentStatus;
}