// ui/src/lib/api/model/deploymentStatus.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */

/**
 * Current status of a deployment
 */
export type DeploymentStatus = typeof DeploymentStatus[keyof typeof DeploymentStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeploymentStatus = {
  StatusActive: 'active',
  StatusInactive: 'inactive',
  StatusDeploying: 'deploying',
  StatusFailed: 'failed',
  StatusDeprecated: 'deprecated',
} as const;